new pipeline, ready to test
